requirements.txt
pandas
numpy
psycopg2-binary
python-dotenv
jupyter